+++ R/dtwclust-plot.R
@@ -43,8 +43,8 @@ NULL
 #' @aliases plot,dtwclust,missing-method
 #'
 setMethod("plot", signature(x="dtwclust", y="missing"),
-          function(x, y, clus=seq_len(x@k),
-                   labs.arg = NULL, data=NULL, time = NULL, plot = TRUE, ...) {
+          function(x, y, clus = seq_len(x@k),
+                   labs.arg = NULL, data = NULL, time = NULL, plot = TRUE, ...) {
 
                ## Obtain data, the priority is: provided data > included data matrix > included data list
 
@@ -82,8 +82,6 @@ setMethod("plot", signature(x="dtwclust", y="missing"),
                          or provide the data manually.")
                }
 
-               df <- as.data.frame(df)
-
                ## Obtain centers (which can be matrix or lists of series)
 
                if (is.matrix(x@centers)) {
@@ -112,9 +110,11 @@ setMethod("plot", signature(x="dtwclust", y="missing"),
 
                ## transform data
 
-               #n <- nrow(df)
+               df <- as.data.frame(df)
+
                if (is.null(time)) {
                     t <- seq_len(L)
+
                } else {
                     if (length(time) != L)
                          stop("Length mismatch between values and time stamps")